Fix deserializeDelegateRegistration()
Transaction.Serialized is a binary []byte, not ASCII hex string.

+++ crypto/fixtures.go
@@ -75,7 +75,7 @@ type TestingSecondSignatureRegistrationFixture struct {
 
 type TestingDelegateRegistrationFixture struct {
 	Data struct {
-		Type            uint8     `json:"type,omitempty"`
+		Type            uint16    `json:"type,omitempty"`
 		Network         uint8     `json:"network,omitempty"`
 		Version         uint8     `json:"version,omitempty"`
 		Amount          FlexToshi `json:"amount,omitempty"`
